+++ common/src/main/java/com/stevekung/fishofthieves/client/model/SplashtailModel.java
@@ -54,7 +54,7 @@ public class SplashtailModel<T extends Splashtail> extends EntityModel<T> implem
         var backRotation = 1.0f;
         var backRotSpeed = 1.0f;
 
-        if (!entity.isInWater())
+        if (!entity.isInWater() && !entity.isNoFlip())
         {
             backRotation = 1.5f;
             backRotSpeed = 1.7f;