+++ common/src/main/java/com/stevekung/fishofthieves/client/model/WildsplashModel.java
@@ -60,7 +60,7 @@ public class WildsplashModel<T extends Wildsplash> extends EntityModel<T> implem
         var backRotation = 1.0f;
         var backRotSpeed = 1.0f;
 
-        if (!entity.isInWater())
+        if (!entity.isInWater() && !entity.isNoFlip())
         {
             backRotation = 1.6f;
             backRotSpeed = 1.8f;