+++ common/src/main/java/com/stevekung/fishofthieves/client/renderer/entity/AncientscaleRenderer.java
@@ -17,14 +17,15 @@ public class AncientscaleRenderer extends ThievesFishRenderer<AncientscaleVarian
     }
 
     @Override
-    protected void setupRotations(Ancientscale ancientscale, PoseStack poseStack, float ageInTicks, float rotationYaw, float partialTicks)
+    protected void setupRotations(Ancientscale entity, PoseStack poseStack, float ageInTicks, float rotationYaw, float partialTicks)
     {
-        super.setupRotations(ancientscale, poseStack, ageInTicks, rotationYaw, partialTicks);
+        super.setupRotations(entity, poseStack, ageInTicks, rotationYaw, partialTicks);
+        var inWater = entity.isInWater() || entity.isNoFlip();
         var bodyRotBase = 1.0f;
-        var baseDegree = ancientscale.isPartying() ? -20.0f : 5.0f;
-        var bodyRotSpeed = ancientscale.isPartying() ? ancientscale.isInWater() ? 2.0f : 1.0f : 0.65f;
+        var baseDegree = entity.isPartying() ? -20.0f : 5.0f;
+        var bodyRotSpeed = entity.isPartying() ? inWater ? 2.0f : 1.0f : 0.65f;
 
-        if (!ancientscale.isInWater())
+        if (!inWater)
         {
             bodyRotBase = 1.7f;
         }
@@ -32,7 +33,7 @@ public class AncientscaleRenderer extends ThievesFishRenderer<AncientscaleVarian
         var degree = baseDegree * Mth.sin(bodyRotBase * bodyRotSpeed * ageInTicks);
         poseStack.mulPose(Vector3f.YP.rotationDegrees(degree));
 
-        if (!ancientscale.isInWater())
+        if (!inWater)
         {
             poseStack.translate(0.165f, 0.1f, 0.1f);
             poseStack.mulPose(Vector3f.ZP.rotationDegrees(90.0f));