+++ common/src/main/java/com/stevekung/fishofthieves/client/renderer/entity/WreckerRenderer.java
@@ -19,14 +19,15 @@ public class WreckerRenderer extends ThievesFishRenderer<WreckerVariant, Wrecker
     }
 
     @Override
-    protected void setupRotations(Wrecker wrecker, PoseStack poseStack, float ageInTicks, float rotationYaw, float partialTicks)
+    protected void setupRotations(Wrecker entity, PoseStack poseStack, float ageInTicks, float rotationYaw, float partialTicks)
     {
-        super.setupRotations(wrecker, poseStack, ageInTicks, rotationYaw, partialTicks);
+        super.setupRotations(entity, poseStack, ageInTicks, rotationYaw, partialTicks);
+        var inWater = entity.isInWater() || entity.isNoFlip();
         var bodyRotBase = 1.0f;
-        var baseDegree = wrecker.isPartying() ? -20.0f : 4.0f;
-        var bodyRotSpeed = wrecker.isPartying() ? wrecker.isInWater() ? 2.0f : 1.0f : 0.6f;
+        var baseDegree = entity.isPartying() ? -20.0f : 4.0f;
+        var bodyRotSpeed = entity.isPartying() ? inWater ? 2.0f : 1.0f : 0.6f;
 
-        if (!wrecker.isInWater())
+        if (!inWater)
         {
             bodyRotBase = 1.7f;
         }
@@ -34,7 +35,7 @@ public class WreckerRenderer extends ThievesFishRenderer<WreckerVariant, Wrecker
         var degree = baseDegree * Mth.sin(bodyRotBase * bodyRotSpeed * ageInTicks);
         poseStack.mulPose(Vector3f.YP.rotationDegrees(degree));
 
-        if (!wrecker.isInWater())
+        if (!inWater)
         {
             poseStack.translate(0.15f, 0.1f, 0.0f);
             poseStack.mulPose(Vector3f.ZP.rotationDegrees(90.0f));