+++ CHANGE_LOG.md
@@ -15,6 +15,7 @@
 * Added Russian translation, Thanks to Telezhka-the-First
 * Added Fish Bone drop to fishes
 * Allow modders to add their own fish variant
+* Added NoFlip Tag
 * Blackcloud Wrecker is now a rare variant
 * Fishes are now follow the players if holding any worms
 * Worms can now feed to fishes, depends on their type
@@ -26,6 +27,7 @@
 * Balanced all chests/entities loot
 * Balanced smelting EXP
 * All models are now using RGBA to the renderer
+* Change Creative Tabs icon to Splashtail
 * Update advancements description
 * Update all fish model for more Minecrafty style
 * Fix "fishes" for config, datapack